requirements.txt
# Run on Raspberry Pi (Python 3)
# Install: pip install -r requirements.txt
spidev
RPi.GPIO
Pillow
